[move] #19 스탬프 api 패키징을 통한 파일 위치 수정
